+++ src/app/components/forms/suggestion-box/terra-suggestion-box.component.ts
@@ -10,11 +10,11 @@ import {
     SimpleChanges
 } from '@angular/core';
 import { TerraSuggestionBoxValueInterface } from './data/terra-suggestion-box.interface';
-import { NG_VALUE_ACCESSOR } from '@angular/forms';
 import {
-    isNull,
-    isNullOrUndefined
-} from 'util';
+    ControlValueAccessor,
+    NG_VALUE_ACCESSOR
+} from '@angular/forms';
+import { isNullOrUndefined } from 'util';
 import { TerraPlacementEnum } from '../../../helpers/enums/terra-placement.enum';
 import { TerraBaseData } from '../../data/terra-base.data';
 
@@ -32,7 +32,7 @@ const MAX_LASTLY_USED_ENTRIES:number = 5;
         }
     ]
 })
-export class TerraSuggestionBoxComponent implements OnInit, OnChanges
+export class TerraSuggestionBoxComponent implements OnInit, OnChanges, ControlValueAccessor
 {
     @Input()
     public inputName:string;
@@ -47,7 +47,7 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
     public inputTooltipText:string;
 
     @Input()
-    public inputTooltipPlacement:TerraPlacementEnum;
+    public inputTooltipPlacement:TerraPlacementEnum = TerraPlacementEnum.TOP;
 
     @Input()
     public inputListBoxValues:Array<TerraSuggestionBoxValueInterface>;
@@ -55,24 +55,31 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
     @Input()
     public inputWithRecentlyUsed:boolean;
 
+    /**
+     * @deprecated since it notifies the user at exactly the same time as ngModelChange <-> onChangeCallback
+     */
     @Output()
     public outputValueChanged:EventEmitter<TerraSuggestionBoxValueInterface> = new EventEmitter<TerraSuggestionBoxValueInterface>();
 
     @Output()
     public outputClicked:EventEmitter<Event> = new EventEmitter<Event>();
 
-    public isValid:boolean;
-    public selectedValue:TerraSuggestionBoxValueInterface;
+    @Output()
+    public textInputValueChange:EventEmitter<string> = new EventEmitter<string>();
+
+    public isValid:boolean = true;
 
     protected displayListBoxValues:Array<TerraSuggestionBoxValueInterface> = [];
-    protected lastSelectedValues:Array<TerraSuggestionBoxValueInterface>;
-    protected listBoxHeadingKey:string;
+    protected lastSelectedValues:Array<TerraSuggestionBoxValueInterface> = [];
+    protected listBoxHeadingKey:string = '';
     protected noEntriesTextKey:string;
-    protected tmpSelectedValue:TerraSuggestionBoxValueInterface;
+    protected _selectedValue:TerraSuggestionBoxValueInterface = null;
+    protected tmpSelectedValue:TerraSuggestionBoxValueInterface = null;
+    protected _textInputValue:string;
+    protected _toggleOpen:boolean = false;
 
-    private _toggleOpen:boolean;
     private hasLabel:boolean;
-    private _value:number | string | TerraBaseData;
+
     private clickListener:(event:Event) => void;
 
     constructor(private elementRef:ElementRef)
@@ -116,16 +123,8 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
             if(changes['inputListBoxValues'].currentValue.length > 0
                    && !this.inputListBoxValues.find((x:TerraSuggestionBoxValueInterface):boolean => this.selectedValue === x))
             {
-                this.select(this.inputListBoxValues[0]);
-            }
-            else
-            {
-                // reset selected value if the list is empty
-                this.selectedValue =
-                    {
-                        value:   '',
-                        caption: ''
-                    };
+                // reset selected value if the value does not exists or the list is empty
+                this.selectedValue = null;
             }
         }
     }
@@ -151,38 +150,24 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
 
     public get value():number | string | TerraBaseData
     {
-        return this._value;
+        return !isNullOrUndefined(this.selectedValue) ? this.selectedValue.value : null;
     }
 
     public set value(value:number | string | TerraBaseData)
     {
-        this._value = value;
-
-        if(!isNullOrUndefined(value))
+        if(isNullOrUndefined(this.inputListBoxValues))
         {
-            let selectedValue:TerraSuggestionBoxValueInterface =
-                this.inputListBoxValues.find((item:TerraSuggestionBoxValueInterface) => item.value === value);
-
-            if(selectedValue)
-            {
-                this.selectedValue = {
-                    caption: selectedValue.caption,
-                    value:   selectedValue.value
-                };
-            }
+            this.selectedValue = null;
         }
-        else if(!isNullOrUndefined(this.inputListBoxValues) &&  this.inputListBoxValues.length)
+        else
         {
-            this.selectedValue = {
-                caption: this.inputListBoxValues[0].caption,
-                value:   this.inputListBoxValues[0].value
-            };
+            this.selectedValue = this.inputListBoxValues.find((item:TerraSuggestionBoxValueInterface) => item.value === value);
         }
 
         this.tmpSelectedValue = this.selectedValue;
     }
 
-    private onClick(evt:Event):void
+    protected onClick(evt:Event):void
     {
         evt.stopPropagation(); // prevents the click listener on the document to be fired right after
         this.toggleOpen = !this.toggleOpen;
@@ -225,10 +210,7 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
         }
 
         // update selected value
-        this.selectedValue = {
-            caption: value.caption,
-            value:   value.value
-        };
+        this.selectedValue = value;
 
         // update last selected values
         if(this.inputWithRecentlyUsed)
@@ -275,29 +257,32 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
 
     public onChange():void
     {
-        let searchString:any = this.selectedValue.caption;
+        let searchString:any = this.textInputValue;
         this.toggleOpen = true;
 
         if(searchString.length >= 3)
         {
             this.listBoxHeadingKey = 'terraSuggestionBox.suggestions';
             this.noEntriesTextKey = 'terraSuggestionBox.noSuggestions';
-            this.displayListBoxValues = this.inputListBoxValues.filter((value:TerraSuggestionBoxValueInterface) =>
+            if(!isNullOrUndefined(this.inputListBoxValues))
             {
-                // check if search string has a full match
-                if(value.caption.toUpperCase().includes(searchString.toUpperCase()))
+                this.displayListBoxValues = this.inputListBoxValues.filter((value:TerraSuggestionBoxValueInterface) =>
                 {
-                    return true;
-                }
+                    // check if search string has a full match
+                    if(value.caption.toUpperCase().includes(searchString.toUpperCase()))
+                    {
+                        return true;
+                    }
 
-                // search for partial strings
-                let searchStringIncluded:boolean = true;
-                searchString.split(' ').forEach((word:string) =>
-                {
-                    searchStringIncluded = searchStringIncluded && value.caption.toUpperCase().includes(word.toUpperCase());
+                    // search for partial strings
+                    let searchStringIncluded:boolean = true;
+                    searchString.split(' ').forEach((word:string) =>
+                    {
+                        searchStringIncluded = searchStringIncluded && value.caption.toUpperCase().includes(word.toUpperCase());
+                    });
+                    return searchStringIncluded;
                 });
-                return searchStringIncluded;
-            });
+            }
         }
         else if(this.inputWithRecentlyUsed)
         {
@@ -305,14 +290,18 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
             this.noEntriesTextKey = 'terraSuggestionBox.noRecentlyUsed';
             this.displayListBoxValues = this.lastSelectedValues;
         }
-        else
+        else if(!isNullOrUndefined(this.inputListBoxValues))
         {
             this.displayListBoxValues = this.inputListBoxValues;
         }
 
-        this.value = this.selectedValue.value;
+        // update selected value
+        this.selectedValue = this.displayListBoxValues.find((val:TerraSuggestionBoxValueInterface) => val.caption === searchString);
     }
 
+    /**
+     * @deprecated use ngModel instead to reset the selected value
+     */
     public resetComponentValue():void
     {
         this.value = null;
@@ -326,7 +315,7 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
         this.tmpSelectedValue = null;
     }
 
-    private onKeyDown(event:KeyboardEvent):void
+    protected onKeyDown(event:KeyboardEvent):void
     {
         // check if one of the dedicated keys has been pressed
         if(!(event.key === 'ArrowDown' || event.key === 'ArrowUp' || event.key === 'Enter' || event.key === 'Escape'))
@@ -420,7 +409,7 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
      * workaround to prevent calling the select() method on the label click
      * @param event
      */
-    private onInputClick(event:any):void
+    protected onInputClick(event:any):void
     {
         this.outputClicked.emit(event);
 
@@ -430,5 +419,49 @@ export class TerraSuggestionBoxComponent implements OnInit, OnChanges
             // select the input text <-> mark all
             event.target.select();
         }
+    }
+
+    protected get textInputValue():string
+    {
+        return this._textInputValue;
+    }
+
+    protected set textInputValue(value:string)
+    {
+        if(this._textInputValue !== value)
+        {
+            this.textInputValueChange.emit(value);
+        }
+        this._textInputValue = value;
+    }
+
+    public set selectedValue(value:TerraSuggestionBoxValueInterface)
+    {
+        // does not do anything if the value changes from undefined to null or reverse
+        if(isNullOrUndefined(this._selectedValue) && isNullOrUndefined(value))
+        {
+            return;
+        }
+        // the value has changed?
+        if(this._selectedValue !== value)
+        {
+            // update local model
+            this._selectedValue = value;
+            this.tmpSelectedValue = this._selectedValue;
+
+            // execute callback functions
+            this.onTouchedCallback(); // this may be called when the text input value changes instead!?
+            this.onChangeCallback(this.value);
+            this.outputValueChanged.emit(this._selectedValue);
+
+            // finally update text input value
+            // This needs to be done after executing the callbacks to make a live search work!!
+            this.textInputValue = !isNullOrUndefined(this._selectedValue) ? this._selectedValue.caption : undefined;
+        }
+    }
+
+    public get selectedValue():TerraSuggestionBoxValueInterface
+    {
+        return this._selectedValue;
     }
 }