+++ src/app/components/nested-data-picker/data/nested-data.interface.ts
@@ -1,0 +1,14 @@
+import { NestedDetailDataInterface } from './nested-detail-data.interface';
+
+/**
+ * @author chirila-ioan-daniel
+ */
+export interface NestedDataInterface<T>
+{
+    id?:number;
+    hasChildren?:boolean;
+    details?:Array<NestedDetailDataInterface>;
+    parentId?:number;
+    isLastPage?:boolean;
+    data?:T;
+}