+++ src/app/terra-components.module.ts
@@ -87,16 +87,15 @@ import { TerraBaseService } from './service/terra-base.service';
 import { TerraUrlParamsDecorator } from './service/data/terra-url-params-decorator.service';
 import { TerraAlertComponent } from './components/alert/terra-alert.component';
 import { TerraTimePickerComponent } from './components/forms/input/time-picker/terra-time-picker.component';
-import { CategoryTreeConfig } from './components/category-picker/config/category-tree.config';
 import { TerraJsonToFormFieldService } from './components/forms/dynamic-form/service/terra-json-to-form-field.service';
 import { TerraDynamicFormComponent } from './components/forms/dynamic-form/terra-dynamic-form.component';
 import { TerraDynamicSwitchComponent } from './components/forms/dynamic-form/dynamic-switch/terra-dynamic-switch.component';
 import { TerraDynamicFormService } from './components/forms/dynamic-form/service/terra-dynamic-form.service';
 import { TerraFormFieldControlService } from './components/forms/dynamic-form/service/terra-form-field-control.service';
 import { TerraMultiCheckBoxComponent } from './components/forms/multi-check-box/terra-multi-check-box.component';
+import { TerraNestedDataPickerComponent } from './components/nested-data-picker/terra-nested-data-picker.component';
 import { TerraFileChooserComponent } from './components/buttons/file-chooser/terra-file-chooser.component';
 import { TerraCodeEditorComponent } from './components/editors/code-editor/terra-code-editor.component';
-// ### import of terra-component examples
 import { TerraButtonComponentExample } from './components/buttons/button/example/terra-button.component.example';
 import { TerraAlertComponentExample } from './components/alert/example/terra-alert.component.example';
 import { TerraCardComponentExample } from './components/layouts/card/example/terra-card.component.example';
@@ -131,8 +130,7 @@ import { TerraNoResultNoticeComponentExample } from './components/no-result/exam
 import { TerraButtonWithOptionsComponentExample } from './components/buttons/button-with-options/example/terra-button-with-options.component.example';
 import { TerraInputComponentExample } from './components/forms/input/example/terra-input.component.example';
 import { TerraFilterComponentExample } from './components/filter/example/terra-filter.component.example';
-import { TerraDataTableContextMenuComponentExample }
-from './components/tables/data-table/context-menu/example/terra-data-table-context-menu.component.example';
+import { TerraDataTableContextMenuComponentExample } from './components/tables/data-table/context-menu/example/terra-data-table-context-menu.component.example';
 import { TerraNoteEditorComponentExample } from './components/editors/note-editor/example/terra-note-editor.component.example';
 import { TerraNoteComponentExample } from './components/note/example/terra-note.component.example';
 import { TerraCodeEditorComponentExample } from './components/editors/code-editor/example/terra-code-editor.component.example';
@@ -145,11 +143,20 @@ import { TerraMultiSplitViewBreadcrumbsService } from './components/split-view/m
 import { RouterModule } from '@angular/router';
 import { TerraTimePickerComponentExample } from './components/forms/input/time-picker/example/terra-time-picker.component.example';
 import { TerraTwoColumnsContainerComponent } from './components/layouts/column-container/terra-two-columns-container.component';
+import { TerraNestedDataPickerComponentExample } from './components/nested-data-picker/example/terra-nested-data-picker.component.example';
+import { TerraFormComponent } from './components/forms/form/terra-form.component';
+import { TerraFormEntryComponent } from './components/forms/form/form-entry/terra-form-entry.component';
+// tslint:disable-next-line
+import { TerraFormEntryListComponent } from './components/forms/form/form-entry-list/terra-form-entry-list.component';
+// tslint:disable-next-line
+import { TerraFormContainerComponent } from './components/forms/form/form-container/terra-form-container.component';
 import { TerraTwoColumnsContainerDirective } from './components/layouts/column-container/terra-two-columns-container.directive';
 import { TerraBreadcrumbsComponent } from './components/breadcrumbs/terra-breadcrumbs.component';
 import { TerraBreadcrumbsService } from './components/breadcrumbs/service/terra-breadcrumbs.service';
 import { Type } from '@angular/core/src/type';
 import { DeviceDetectorModule } from 'ngx-device-detector';
+import { TerraTagSelectComponent } from './components/forms/tag-select/terra-tag-select.component';
+import { TerraTagSelectComponentExample } from './components/forms/tag-select/example/terra-tag-select.component.example';
 
 
 function createCompiler(compilerFactory:CompilerFactory):Compiler
@@ -217,10 +224,16 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraCategoryPickerComponent,
         TerraDynamicFormComponent,
         TerraDynamicSwitchComponent,
+        TerraNestedDataPickerComponent,
         TerraCategoryPickerComponent,
         TerraTwoColumnsContainerComponent,
+        TerraFormComponent,
+        TerraFormContainerComponent,
+        TerraFormEntryComponent,
+        TerraFormEntryListComponent,
         TerraTwoColumnsContainerDirective,
         TerraBreadcrumbsComponent,
+        TerraTagSelectComponent,
 
         // ### declarations of terra-components examples
         TerraButtonComponentExample,
@@ -263,7 +276,9 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraCodeEditorComponentExample,
         TerraNoteComponentExample,
         TerraNodeTreeComponentExample,
-        TerraTimePickerComponentExample
+        TerraTimePickerComponentExample,
+        TerraNestedDataPickerComponentExample,
+        TerraTagSelectComponentExample
     ],
     entryComponents: [
         TerraAlertPanelComponent,
@@ -316,8 +331,14 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraNodeTreeComponent,
         TerraNodeComponent,
         TerraCategoryPickerComponent,
+        TerraNestedDataPickerComponent,
         TerraTwoColumnsContainerComponent,
-        TerraBreadcrumbsComponent
+        TerraFormComponent,
+        TerraFormContainerComponent,
+        TerraFormEntryComponent,
+        TerraFormEntryListComponent,
+        TerraBreadcrumbsComponent,
+        TerraTagSelectComponent
     ],
     exports:         [
         TerraAlertPanelComponent,
@@ -376,10 +397,16 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraCategoryPickerComponent,
         TerraDynamicFormComponent,
         TerraDynamicSwitchComponent,
+        TerraNestedDataPickerComponent,
         TerraCategoryPickerComponent,
         TerraTwoColumnsContainerComponent,
+        TerraFormComponent,
+        TerraFormContainerComponent,
+        TerraFormEntryComponent,
+        TerraFormEntryListComponent,
         TerraTwoColumnsContainerDirective,
         TerraBreadcrumbsComponent,
+        TerraTagSelectComponent,
 
         // ### declarations of terra-components examples
         TerraButtonComponentExample,
@@ -424,7 +451,9 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraDynamicFormComponent,
         TerraDynamicSwitchComponent,
         TerraNodeTreeComponentExample,
-        TerraTimePickerComponentExample
+        TerraTimePickerComponentExample,
+        TerraNestedDataPickerComponentExample,
+        TerraTagSelectComponentExample
     ],
     imports:         [
         BrowserModule,